# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /generations:
    get:
      operationId: listGenerations
      tags: [generations]
      summary: List generation history
      parameters:
        - in: query
          name: page
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: pageSize
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 24 }
      responses:
        "200":
          description: Paginated generation history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedGenerations"
    post:
      operationId: createGeneration
      tags: [generations]
      summary: Submit a generation job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerationInput"
      responses:
        "201":
          description: Queued generation job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationJob"
  /generations/{id}:
    get:
      operationId: getGeneration
      tags: [generations]
      summary: Get a generation job
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Generation job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationJob"
    delete:
      operationId: deleteGeneration
      tags: [generations]
      summary: Delete a completed generation from history
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Deleted
        "404":
          description: Generation not found
        "409":
          description: Active generations must be cancelled before deletion
  /generations/{id}/cancel:
    post:
      operationId: cancelGeneration
      tags: [generations]
      summary: Cancel an active generation job
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Cancelled generation job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationJob"
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    GenerationInput:
      type: object
      required: [prompt, generationMode, durationSeconds, fps, width, height, qualityPreset, seedMode]
      properties:
        provider: { type: string, enum: [COMFYUI, FAL], default: COMFYUI }
        model:
          type: string
          enum: [veo-3.1-fast, kling-v3-standard, seedance-2.0-mini, seedance-2.0]
        voiceCloningEnabled: { type: boolean, default: false }
        characterIds: { type: array, items: { type: string }, minItems: 1, maxItems: 9 }
        settingId: { type: string }
        prompt: { type: string, minLength: 1, maxLength: 10000 }
        negativePrompt: { type: string, maxLength: 5000 }
        cameraInstructions: { type: string, maxLength: 5000 }
        dialogue: { type: string, maxLength: 5000 }
        motionInstructions: { type: string, maxLength: 5000 }
        audioInstructions: { type: string, maxLength: 5000 }
        generationMode: { type: string }
        referenceVideoKey: { type: string }
        durationSeconds: { type: number, minimum: 1, maximum: 120 }
        fps: { type: number, enum: [24, 25, 30] }
        width: { type: number, minimum: 64, maximum: 4096 }
        height: { type: number, minimum: 64, maximum: 4096 }
        qualityPreset: { type: string, enum: [DRAFT, STANDARD, HIGH] }
        seedMode: { type: string, enum: [RANDOM, FIXED] }
        seed: { type: ["number", "null"], minimum: 0 }
    GenerationJob:
      type: object
      required: [id, title, status, provider, providerModelId, providerRequestId, providerTaskMetadata, voiceCloningEnabled, prompt, compiledPrompt, generationMode, qualityPreset, width, height, fps, durationSeconds, progress, currentNode, serverName, workflowName, longFormProjectId, longFormProjectTitle, longFormSceneNumber, longFormShotNumber, createdAt]
      properties:
        id: { type: string }
        title: { type: string }
        status: { type: string, enum: [DRAFT, UPLOADING, QUEUED, RUNNING, DOWNLOADING, COMPLETED, FAILED, CANCELLED] }
        provider: { type: string, enum: [COMFYUI, FAL] }
        providerModelId: { type: ["string", "null"] }
        providerRequestId: { type: ["string", "null"] }
        providerTaskMetadata:
          type: object
          additionalProperties: true
        voiceCloningEnabled: { type: boolean }
        prompt: { type: string }
        compiledPrompt: { type: string }
        generationMode: { type: string }
        qualityPreset: { type: string }
        width: { type: number }
        height: { type: number }
        fps: { type: number }
        durationSeconds: { type: number }
        seed: { type: ["number", "null"] }
        progress: { type: ["number", "null"] }
        currentNode: { type: ["string", "null"] }
        serverName: { type: ["string", "null"] }
        workflowName: { type: ["string", "null"] }
        longFormProjectId: { type: ["string", "null"] }
        longFormProjectTitle: { type: ["string", "null"] }
        longFormSceneNumber: { type: ["number", "null"] }
        longFormShotNumber: { type: ["number", "null"] }
        comfyPromptId: { type: ["string", "null"] }
        outputUrl: { type: ["string", "null"] }
        errorMessage: { type: ["string", "null"] }
        createdAt: { type: string }
        queuedAt: { type: ["string", "null"] }
        completedAt: { type: ["string", "null"] }
    PaginatedGenerations:
      type: object
      required: [items, page, pageSize, totalItems, totalPages]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/GenerationJob"
        page: { type: number }
        pageSize: { type: number }
        totalItems: { type: number }
        totalPages: { type: number }
...